move routes into root app dir
