--- app/(views)/page.tsx
@@ -1,5 +1,0 @@
-import { BlankPage } from '@/components/pages/BlankPage'
-
-export default function Page() {
-  return <BlankPage />
-}
