+++ app/chat/[thread_id]/layout.tsx
