+++ app/image/[image_id]/page.tsx
