+++ app/images/[thread_id]/@modal/(...)image/[image_id]/Modal.tsx
