+++ app/images/[thread_id]/@modal/(...)image/[image_id]/default.tsx
