+++ app/images/[thread_id]/@modal/default.tsx
