+++ app/images/[thread_id]/ImagesFeed.tsx
