+++ app/images/[thread_id]/default.tsx
