+++ app/images/[thread_id]/layout.tsx
